Classification
11.24 is a demo of classification of financial data. Scatterplots of
data are observed to be normally distributed and an appropriate
classification rule is applied. First, we look at only the first two
variables. The priors in the first model are set to be equal. The
priors in the second model are 0.95 and 0.05. In the third (equal
priors) and fourth model (p1=0.95 and p2=0.05), we look at all
variables and perform classification. APERs are compared.

In 11.28, we look at the iris dataset and perform classification on
sepal vs petal data.

In 5(b), we run a simulation to obtain a dataset with 3 variables from
MVN and 1 variable from Cauchy distribution and perform classification-
(Add more later)